fix: Add check for undefined values
Check undefined values for substring in spotlightcard
